Refactor: editor 디렉토리 구조 변경
editor 해더관련 컴포넌트를 editor/header 디렉토리로 이동
이동된 디렉토리를 반영한 이미지 경로 수정
불필요한 컴포넌트 코드 삭제
