+++ client/src/components/editor/header/DocumentTitle.jsx
