+++ client/src/components/editor/header/Logo.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import logoImagePath from "../../../public/logo1.png";
+import logoImagePath from "../../../../public/logo1.png";
 
 const HeaderLogoImage = () => {
   return <img alt="logo" src={logoImagePath} />;
