+++ client/src/components/editor/header/TerminalImage.jsx
